alias/functions: make the 'last-week' function use the git author date
Due to the command complexity, the alias was removed and replaced with a function.  The new function
collects all of the commits by the author for the last week, then filters those by author date.  It then
prints the log for those commits, and deduplicates based on the commit subject.  This complexity is due
to git log only limiting by commit date, not author date, with no option to limit by author date available.