git-rprompt: run the git commands with '--no-optional-locks'
A git status command can update the index, which may create a repository
lock.  For purposes of the rprompt, it is less important to be
performant and fully accurate over locking the git repo, which can cause
multiple git commands which are run in rapid succession to fail on the
second invocation (due to the repo lock).